+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 edbee.lib:
 
+- hack #127, Option to show Unicode BIDI characters (CVE-2021-425740). (as red-alert icons for now)
 - fix #125, CMake 3.20 Deprecation Warning
 - fix #123, Clicking on TextMarginComponet now moves focus to TextEditorComponent
 - fix #121, Standard Keys not working correctly (Fixed by fetching ALL keybindings via QKeySequence::keyBinding)